# Conflicts:
#	resources/views/citas/edit.blade.php
#	routes/web.php

+++ app/Http/Controllers/EnfermedadController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Especialidad;
+use App\Paciente;
 use Illuminate\Http\Request;
 
 use App\Enfermedad;
@@ -21,7 +22,7 @@ class EnfermedadController extends Controller
      */
     public function index()
     {
-        $enfermedades = Enfermedad::all();
+        $enfermedades = Enfermedad::all()->sortBy('especialidad_id');
 
         return view('enfermedades/index',['enfermedades'=>$enfermedades]);
 
@@ -124,9 +125,21 @@ class EnfermedadController extends Controller
     public function destroy($id)
     {
         $enfermedad = Enfermedad::find($id);
-        $enfermedad->delete();
-        flash('Enfermedad borrada correctamente');
+
+        $paciente = Paciente::where('enfermedad_id','=',$id)->value('enfermedad_id');
+
+        if($paciente){
+
+            flash('Existe algun paciente con esta enfermedad. Revise los datos.');
+
+        }else{
+
+            $enfermedad->delete();
+            flash('Enfermedad borrada correctamente');
+        }
+
 
         return redirect()->route('enfermedades.index');
+
     }
 }
